--- src/utils/clampValue.ts
@@ -1,3 +1,0 @@
-export const clampValue = (value: number, min: number, max: number): number => {
-  return Math.max(min, Math.min(value, max));
-};
